Working basic denoising CVAE on MNIST data
This is the simplest way I could come up with for implementing
a CVAE with different input and 'target' images.

I'm really not a fan of the current keras example, it seems
to add unnecessary complexity.

It does ok, but note that a small CVAE isn't particularly well-suited
to producing good reconstructions of MNIST digits. You can get
improvements by increasing the number of latent
dimensions and/or lowering `beta` (which reduces the 'smoothness'
enforced on the latent space). It is also important to let it train for
long enough (until the validation loss stops improving). This can take a
while, especially on a CPU.