Working basic denoising conditional CVAE on MNIST data
Based on the example Larissa found together with my intuitions.
Again, decent examples seem to be hard to find online.

I've extended the previous CVAE example to add in a condition: the
label of the digit. I have included some comments in the code
explaining my thought process, and how this MNIST example relates to
our real goals.

This is my first go at this, but it seems to work (based on very
limited testing!).